modularization of iris-mpc-hnsw
