+++ Dockerfile.hawk
@@ -29,7 +29,7 @@ FROM build-image AS build-app
 WORKDIR /src/iris-mpc
 COPY . .
 
-RUN cargo build --release --bin server-hawk
+RUN cargo build --release --bin iris-mpc-hawk
 
 FROM ubuntu:22.04
 ENV DEBIAN_FRONTEND=noninteractive
@@ -38,8 +38,8 @@ RUN apt-get update && apt-get install -y ca-certificates awscli
 COPY certs /usr/local/share/ca-certificates/
 RUN update-ca-certificates
 
-COPY --from=build-app /src/iris-mpc/target/release/server-hawk /bin/server-hawk
+COPY --from=build-app /src/iris-mpc/target/release/iris-mpc-gpu /bin/iris-mpc-gpu
 
 USER 65534
 
-ENTRYPOINT ["/bin/server-hawk"]
+ENTRYPOINT ["/bin/iris-mpc-gpu"]
