+++ iris-mpc/Cargo.toml
@@ -47,15 +47,16 @@ default = []
 
 
 [[bin]]
-name = "server-hawk"
-path = "bin/server_hawk.rs"
+name = "iris-mpc-gpu"
+path = "bin/server.rs"
+
+
+[[bin]]
+name = "iris-mpc-hawk"
+path = "bin/server/iris_mpc_hawk.rs"
 
 
 [[bin]]
 name = "client"
-path = "bin/client.rs"
+path = "bin/client/client.rs"
 
-
-[[bin]]
-name = "server"
-path = "bin/server.rs"
