+++ iris-mpc/src/server/mod.rs
@@ -1,704 +1,51 @@
-#![allow(clippy::needless_range_loop)]
+mod utils;
 
-use aws_sdk_s3::Client as S3Client;
-use aws_sdk_sns::{types::MessageAttributeValue, Client as SNSClient};
-use aws_sdk_sqs::Client;
-use axum::{response::IntoResponse, routing::get, Router};
-use clap::Parser;
-use eyre::{eyre, Context, Report};
-use futures::{stream::BoxStream, StreamExt};
-use iris_mpc::aws::clients::AwsClients;
-use iris_mpc::aws::sns::{send_error_results_to_sns, send_results_to_sns};
-use iris_mpc::init::{initialize_chacha_seeds, initialize_tracing};
-use iris_mpc_common::{
-    config::{Config, ModeOfCompute, Opt},
-    galois_engine::degree4::{GaloisRingIrisCodeShare, GaloisRingTrimmedMaskCodeShare},
-    helpers::{
-        aws::{SPAN_ID_MESSAGE_ATTRIBUTE_NAME, TRACE_ID_MESSAGE_ATTRIBUTE_NAME},
-        inmemory_store::InMemoryStore,
-        key_pair::SharesEncryptionKeyPairs,
-        shutdown_handler::ShutdownHandler,
-        smpc_request::{
-            decrypt_iris_share, get_iris_data_by_party_id, validate_iris_share,
-            CircuitBreakerRequest, IdentityDeletionRequest, ReAuthRequest, ReceiveRequestError,
-            SQSMessage, UniquenessRequest, ANONYMIZED_STATISTICS_MESSAGE_TYPE,
-            CIRCUIT_BREAKER_MESSAGE_TYPE, IDENTITY_DELETION_MESSAGE_TYPE, REAUTH_MESSAGE_TYPE,
-            UNIQUENESS_MESSAGE_TYPE,
-        },
-        smpc_response::{
-            create_message_type_attribute_map, IdentityDeletionResult, ReAuthResult,
-            UniquenessResult, ERROR_FAILED_TO_PROCESS_IRIS_SHARES,
-            ERROR_SKIPPED_REQUEST_PREVIOUS_NODE_BATCH, SMPC_MESSAGE_TYPE_ATTRIBUTE,
-        },
-        sync::{SyncResult, SyncState},
-        task_monitor::TaskMonitor,
-    },
-    iris_db::get_dummy_shares_for_deletion,
-    job::{BatchMetadata, BatchQuery, JobSubmissionHandle},
+use crate::server::utils::get_check_addresses;
+use crate::services::aws::clients::AwsClients;
+use crate::services::aws::sns::send_results_to_sns;
+use crate::services::init::initialize_chacha_seeds;
+use crate::services::processors::batch::receive_batch;
+use crate::services::processors::process_identity_deletions;
+use crate::services::store::load_db;
+use axum::http::StatusCode;
+use axum::response::IntoResponse;
+use axum::routing::get;
+use axum::Router;
+use eyre::{eyre, WrapErr};
+use iris_mpc_common::config::{Config, ModeOfCompute};
+use iris_mpc_common::helpers::inmemory_store::InMemoryStore;
+use iris_mpc_common::helpers::key_pair::SharesEncryptionKeyPairs;
+use iris_mpc_common::helpers::shutdown_handler::ShutdownHandler;
+use iris_mpc_common::helpers::smpc_request::{
+    ANONYMIZED_STATISTICS_MESSAGE_TYPE, IDENTITY_DELETION_MESSAGE_TYPE, REAUTH_MESSAGE_TYPE,
+    UNIQUENESS_MESSAGE_TYPE,
 };
+use iris_mpc_common::helpers::smpc_response::{
+    create_message_type_attribute_map, IdentityDeletionResult, ReAuthResult, UniquenessResult,
+};
+use iris_mpc_common::helpers::sync::{SyncResult, SyncState};
+use iris_mpc_common::helpers::task_monitor::TaskMonitor;
+use iris_mpc_common::iris_db::get_dummy_shares_for_deletion;
+use iris_mpc_common::job::JobSubmissionHandle;
 use iris_mpc_cpu::execution::hawk_main::{
     GraphStore, HawkActor, HawkArgs, HawkHandle, ServerJobResult,
 };
-use iris_mpc_store::{
-    fetch_and_parse_chunks, last_snapshot_timestamp, DbStoredIris, ObjectStore, S3Store,
-    S3StoredIris, Store, StoredIrisRef,
-};
-use reqwest::StatusCode;
+use iris_mpc_store::{S3Store, Store, StoredIrisRef};
 use serde::{Deserialize, Serialize};
-use std::{
-    collections::{HashMap, HashSet},
-    fmt::Debug,
-    mem, panic,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc, LazyLock, Mutex,
-    },
-    time::{Duration, Instant},
-};
-use tokio::{
-    sync::{mpsc, oneshot, Semaphore},
-    task::{spawn_blocking, JoinHandle},
-    time::timeout,
-};
+use std::mem;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::{Arc, LazyLock, Mutex};
+use std::time::{Duration, Instant};
+use tokio::sync::{mpsc, oneshot};
+use tokio::time::timeout;
 
 const RNG_SEED_INIT_DB: u64 = 42;
-const SQS_POLLING_INTERVAL: Duration = Duration::from_secs(1);
-const MAX_CONCURRENT_REQUESTS: usize = 32;
 
-static CURRENT_BATCH_SIZE: LazyLock<Mutex<usize>> = LazyLock::new(|| Mutex::new(0));
+pub const SQS_POLLING_INTERVAL: Duration = Duration::from_secs(1);
+pub const MAX_CONCURRENT_REQUESTS: usize = 32;
+pub static CURRENT_BATCH_SIZE: LazyLock<Mutex<usize>> = LazyLock::new(|| Mutex::new(0));
 
-type GaloisShares = (
-    GaloisRingIrisCodeShare,
-    GaloisRingTrimmedMaskCodeShare,
-    Vec<GaloisRingIrisCodeShare>,
-    Vec<GaloisRingTrimmedMaskCodeShare>,
-    Vec<GaloisRingIrisCodeShare>,
-    Vec<GaloisRingTrimmedMaskCodeShare>,
-);
-type ParseSharesTaskResult = Result<(GaloisShares, GaloisShares), Report>;
-
-fn decode_iris_message_shares(
-    code_share: String,
-    mask_share: String,
-) -> eyre::Result<(GaloisRingIrisCodeShare, GaloisRingTrimmedMaskCodeShare)> {
-    let iris_share = GaloisRingIrisCodeShare::from_base64(&code_share)
-        .context("Failed to base64 parse iris code")?;
-    let mask_share: GaloisRingTrimmedMaskCodeShare =
-        GaloisRingIrisCodeShare::from_base64(&mask_share)
-            .context("Failed to base64 parse iris mask")?
-            .into();
-
-    Ok((iris_share, mask_share))
-}
-
-fn preprocess_iris_message_shares(
-    code_share: GaloisRingIrisCodeShare,
-    mask_share: GaloisRingTrimmedMaskCodeShare,
-) -> eyre::Result<GaloisShares> {
-    let mut code_share = code_share;
-    let mut mask_share = mask_share;
-
-    // Original for storage.
-    let store_iris_shares = code_share.clone();
-    let store_mask_shares = mask_share.clone();
-
-    // With rotations for in-memory database.
-    let db_iris_shares = code_share.all_rotations();
-    let db_mask_shares = mask_share.all_rotations();
-
-    // With Lagrange interpolation.
-    GaloisRingIrisCodeShare::preprocess_iris_code_query_share(&mut code_share);
-    GaloisRingTrimmedMaskCodeShare::preprocess_mask_code_query_share(&mut mask_share);
-
-    Ok((
-        store_iris_shares,
-        store_mask_shares,
-        db_iris_shares,
-        db_mask_shares,
-        code_share.all_rotations(),
-        mask_share.all_rotations(),
-    ))
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn receive_batch(
-    party_id: usize,
-    client: &Client,
-    sns_client: &SNSClient,
-    s3_client: &S3Client,
-    config: &Config,
-    store: &Store,
-    skip_request_ids: &[String],
-    shares_encryption_key_pairs: SharesEncryptionKeyPairs,
-    shutdown_handler: &ShutdownHandler,
-    uniqueness_error_result_attributes: &HashMap<String, MessageAttributeValue>,
-    reauth_error_result_attributes: &HashMap<String, MessageAttributeValue>,
-) -> eyre::Result<Option<BatchQuery>, ReceiveRequestError> {
-    let max_batch_size = config.clone().max_batch_size;
-    let queue_url = &config.clone().requests_queue_url;
-    if shutdown_handler.is_shutting_down() {
-        tracing::info!("Stopping batch receive due to shutdown signal...");
-        return Ok(None);
-    }
-
-    let mut batch_query = BatchQuery::default();
-
-    let semaphore = Arc::new(Semaphore::new(MAX_CONCURRENT_REQUESTS));
-    let mut handles = vec![];
-    let mut msg_counter = 0;
-
-    while msg_counter < *CURRENT_BATCH_SIZE.lock().unwrap() {
-        let rcv_message_output = client
-            .receive_message()
-            .max_number_of_messages(1)
-            .queue_url(queue_url)
-            .send()
-            .await
-            .map_err(ReceiveRequestError::FailedToReadFromSQS)?;
-
-        if let Some(messages) = rcv_message_output.messages {
-            for sqs_message in messages {
-                let message: SQSMessage = serde_json::from_str(sqs_message.body().unwrap())
-                    .map_err(|e| ReceiveRequestError::json_parse_error("SQS body", e))?;
-
-                // messages arrive to SQS through SNS. So, all the attributes set in SNS are
-                // moved into the SQS body.
-                let message_attributes = message.message_attributes;
-
-                let mut batch_metadata = BatchMetadata::default();
-
-                if let Some(trace_id) = message_attributes.get(TRACE_ID_MESSAGE_ATTRIBUTE_NAME) {
-                    let trace_id = trace_id.string_value().unwrap();
-                    batch_metadata.trace_id = trace_id.to_string();
-                }
-                if let Some(span_id) = message_attributes.get(SPAN_ID_MESSAGE_ATTRIBUTE_NAME) {
-                    let span_id = span_id.string_value().unwrap();
-                    batch_metadata.span_id = span_id.to_string();
-                }
-
-                let request_type = message_attributes
-                    .get(SMPC_MESSAGE_TYPE_ATTRIBUTE)
-                    .ok_or(ReceiveRequestError::NoMessageTypeAttribute)?
-                    .string_value()
-                    .ok_or(ReceiveRequestError::NoMessageTypeAttribute)?;
-
-                match request_type {
-                    CIRCUIT_BREAKER_MESSAGE_TYPE => {
-                        let circuit_breaker_request: CircuitBreakerRequest =
-                            serde_json::from_str(&message.message).map_err(|e| {
-                                ReceiveRequestError::json_parse_error("circuit_breaker_request", e)
-                            })?;
-                        metrics::counter!("request.received", "type" => "circuit_breaker")
-                            .increment(1);
-                        client
-                            .delete_message()
-                            .queue_url(queue_url)
-                            .receipt_handle(sqs_message.receipt_handle.unwrap())
-                            .send()
-                            .await
-                            .map_err(ReceiveRequestError::FailedToDeleteFromSQS)?;
-                        if let Some(batch_size) = circuit_breaker_request.batch_size {
-                            // Updating the batch size to ensure we process the messages in the next
-                            // loop
-                            *CURRENT_BATCH_SIZE.lock().unwrap() =
-                                batch_size.clamp(1, max_batch_size);
-                            tracing::info!(
-                                "Updating batch size to {} due to circuit breaker message",
-                                batch_size
-                            );
-                        }
-                    }
-
-                    IDENTITY_DELETION_MESSAGE_TYPE => {
-                        // If it's a deletion request, we just store the serial_id and continue.
-                        // Deletion will take place when batch process starts.
-                        let identity_deletion_request: IdentityDeletionRequest =
-                            serde_json::from_str(&message.message).map_err(|e| {
-                                ReceiveRequestError::json_parse_error(
-                                    "Identity deletion request",
-                                    e,
-                                )
-                            })?;
-                        metrics::counter!("request.received", "type" => "identity_deletion")
-                            .increment(1);
-                        batch_query
-                            .deletion_requests_indices
-                            .push(identity_deletion_request.serial_id - 1); // serial_id is 1-indexed
-                        batch_query.deletion_requests_metadata.push(batch_metadata);
-                        client
-                            .delete_message()
-                            .queue_url(queue_url)
-                            .receipt_handle(sqs_message.receipt_handle.unwrap())
-                            .send()
-                            .await
-                            .map_err(ReceiveRequestError::FailedToDeleteFromSQS)?;
-                    }
-
-                    UNIQUENESS_MESSAGE_TYPE => {
-                        msg_counter += 1;
-
-                        let shares_encryption_key_pairs = shares_encryption_key_pairs.clone();
-
-                        let uniqueness_request: UniquenessRequest =
-                            serde_json::from_str(&message.message).map_err(|e| {
-                                ReceiveRequestError::json_parse_error("Uniqueness request", e)
-                            })?;
-                        metrics::counter!("request.received", "type" => "uniqueness_verification")
-                            .increment(1);
-                        store
-                            .mark_requests_deleted(&[uniqueness_request.signup_id.clone()])
-                            .await
-                            .map_err(ReceiveRequestError::FailedToMarkRequestAsDeleted)?;
-
-                        client
-                            .delete_message()
-                            .queue_url(queue_url)
-                            .receipt_handle(sqs_message.receipt_handle.unwrap())
-                            .send()
-                            .await
-                            .map_err(ReceiveRequestError::FailedToDeleteFromSQS)?;
-
-                        if skip_request_ids.contains(&uniqueness_request.signup_id) {
-                            // Some party (maybe us) already meant to delete this request, so we
-                            // skip it. Ignore this message when calculating the batch size.
-                            msg_counter -= 1;
-                            metrics::counter!("skip.request.deleted.sqs.request").increment(1);
-                            tracing::warn!(
-                                "Skipping request due to it being from synced deleted ids: {}",
-                                uniqueness_request.signup_id
-                            );
-                            let message = UniquenessResult::new_error_result(
-                                config.party_id,
-                                uniqueness_request.signup_id,
-                                ERROR_SKIPPED_REQUEST_PREVIOUS_NODE_BATCH,
-                            );
-                            // shares
-                            send_error_results_to_sns(
-                                serde_json::to_string(&message).unwrap(),
-                                &batch_metadata,
-                                sns_client,
-                                config,
-                                uniqueness_error_result_attributes,
-                                UNIQUENESS_MESSAGE_TYPE,
-                            )
-                            .await?;
-                            continue;
-                        }
-
-                        if let Some(batch_size) = uniqueness_request.batch_size {
-                            // Updating the batch size instantly makes it a bit unpredictable, since
-                            // if we're already above the new limit, we'll still process the current
-                            // batch at the higher limit. On the other
-                            // hand, updating it after the batch is
-                            // processed would not let us "unblock" the protocol if we're stuck with
-                            // low throughput.
-                            *CURRENT_BATCH_SIZE.lock().unwrap() =
-                                batch_size.clamp(1, max_batch_size);
-                            tracing::info!("Updating batch size to {}", batch_size);
-                        }
-                        if config.luc_enabled {
-                            if config.luc_lookback_records > 0 {
-                                batch_query.luc_lookback_records = config.luc_lookback_records;
-                            }
-                            if config.luc_serial_ids_from_smpc_request {
-                                if let Some(serial_ids) =
-                                    uniqueness_request.or_rule_serial_ids.clone()
-                                {
-                                    // convert from 1-based serial id to 0-based index in actor
-                                    batch_query
-                                        .or_rule_indices
-                                        .push(serial_ids.iter().map(|x| x - 1).collect());
-                                } else {
-                                    tracing::error!(
-                                        "Received a uniqueness request without serial_ids"
-                                    );
-                                }
-                            }
-                        }
-
-                        batch_query
-                            .request_ids
-                            .push(uniqueness_request.signup_id.clone());
-                        batch_query
-                            .request_types
-                            .push(UNIQUENESS_MESSAGE_TYPE.to_string());
-                        batch_query.metadata.push(batch_metadata);
-
-                        let semaphore = Arc::clone(&semaphore);
-                        let s3_client_arc = s3_client.clone();
-                        let bucket_name = config.shares_bucket_name.clone();
-                        let s3_key = uniqueness_request.s3_key.clone();
-                        let handle = get_iris_shares_parse_task(
-                            party_id,
-                            shares_encryption_key_pairs,
-                            semaphore,
-                            s3_client_arc,
-                            bucket_name,
-                            s3_key,
-                        )?;
-
-                        handles.push(handle);
-                    }
-
-                    REAUTH_MESSAGE_TYPE => {
-                        let shares_encryption_key_pairs = shares_encryption_key_pairs.clone();
-
-                        let reauth_request: ReAuthRequest = serde_json::from_str(&message.message)
-                            .map_err(|e| {
-                                ReceiveRequestError::json_parse_error("Reauth request", e)
-                            })?;
-                        metrics::counter!("request.received", "type" => "reauth").increment(1);
-
-                        tracing::debug!("Received reauth request: {:?}", reauth_request);
-
-                        // TODO: populate sync mechanism table (TBD: rollback or rollforward)
-
-                        client
-                            .delete_message()
-                            .queue_url(queue_url)
-                            .receipt_handle(sqs_message.receipt_handle.unwrap())
-                            .send()
-                            .await
-                            .map_err(ReceiveRequestError::FailedToDeleteFromSQS)?;
-
-                        if reauth_request.use_or_rule
-                            && !(config.luc_enabled && config.luc_serial_ids_from_smpc_request)
-                        {
-                            tracing::error!(
-                                "Received a reauth request with use_or_rule set to true, but LUC \
-                                 is not enabled. Skipping request."
-                            );
-                            continue;
-                        }
-
-                        if config.enable_reauth {
-                            msg_counter += 1;
-
-                            if let Some(batch_size) = reauth_request.batch_size {
-                                // Updating the batch size instantly makes it a bit unpredictable,
-                                // since if we're already above the
-                                // new limit, we'll still process the current
-                                // batch at the higher limit. On the other
-                                // hand, updating it after the batch is
-                                // processed would not let us "unblock" the protocol if we're stuck
-                                // with low throughput.
-                                *CURRENT_BATCH_SIZE.lock().unwrap() =
-                                    batch_size.clamp(1, max_batch_size);
-                                tracing::info!("Updating batch size to {}", batch_size);
-                            }
-
-                            batch_query
-                                .request_ids
-                                .push(reauth_request.reauth_id.clone());
-                            batch_query
-                                .request_types
-                                .push(REAUTH_MESSAGE_TYPE.to_string());
-                            batch_query.metadata.push(batch_metadata);
-                            batch_query.reauth_target_indices.insert(
-                                reauth_request.reauth_id.clone(),
-                                reauth_request.serial_id - 1,
-                            );
-                            batch_query.reauth_use_or_rule.insert(
-                                reauth_request.reauth_id.clone(),
-                                reauth_request.use_or_rule,
-                            );
-
-                            let or_rule_indices = if reauth_request.use_or_rule {
-                                vec![reauth_request.serial_id - 1]
-                            } else {
-                                vec![]
-                            };
-                            batch_query.or_rule_indices.push(or_rule_indices);
-
-                            let semaphore = Arc::clone(&semaphore);
-                            let s3_client_clone = s3_client.clone();
-                            let bucket_name = config.shares_bucket_name.clone();
-                            let s3_key = reauth_request.s3_key.clone();
-                            let handle = get_iris_shares_parse_task(
-                                party_id,
-                                shares_encryption_key_pairs,
-                                semaphore,
-                                s3_client_clone,
-                                bucket_name,
-                                s3_key,
-                            )?;
-
-                            handles.push(handle);
-                        } else {
-                            tracing::warn!("Reauth is disabled, skipping reauth request");
-                        }
-                    }
-
-                    _ => {
-                        client
-                            .delete_message()
-                            .queue_url(queue_url)
-                            .receipt_handle(sqs_message.receipt_handle.unwrap())
-                            .send()
-                            .await
-                            .map_err(ReceiveRequestError::FailedToDeleteFromSQS)?;
-                        tracing::error!("Error: {}", ReceiveRequestError::InvalidMessageType);
-                    }
-                }
-            }
-        } else {
-            tokio::time::sleep(SQS_POLLING_INTERVAL).await;
-        }
-    }
-    for (index, handle) in handles.into_iter().enumerate() {
-        let (
-            (
-                (
-                    store_iris_shares_left,
-                    store_mask_shares_left,
-                    db_iris_shares_left,
-                    db_mask_shares_left,
-                    iris_shares_left,
-                    mask_shares_left,
-                ),
-                (
-                    store_iris_shares_right,
-                    store_mask_shares_right,
-                    db_iris_shares_right,
-                    db_mask_shares_right,
-                    iris_shares_right,
-                    mask_shares_right,
-                ),
-            ),
-            valid_entry,
-        ) = match handle
-            .await
-            .map_err(ReceiveRequestError::FailedToJoinHandle)?
-        {
-            Ok(res) => (res, true),
-            Err(e) => {
-                tracing::error!("Failed to process iris shares: {:?}", e);
-                // Return error message back to the signup-service if failed to process iris
-                // shares
-                let request_id = batch_query.request_ids[index].clone();
-                let (result_attributes, message) = match batch_query.request_types[index].as_str() {
-                    UNIQUENESS_MESSAGE_TYPE => {
-                        let message = UniquenessResult::new_error_result(
-                            config.party_id,
-                            request_id,
-                            ERROR_FAILED_TO_PROCESS_IRIS_SHARES,
-                        );
-                        let serialized = serde_json::to_string(&message).unwrap();
-                        (uniqueness_error_result_attributes, serialized)
-                    }
-                    REAUTH_MESSAGE_TYPE => {
-                        let message = ReAuthResult::new_error_result(
-                            request_id.clone(),
-                            config.party_id,
-                            *batch_query.reauth_target_indices.get(&request_id).unwrap(),
-                            ERROR_FAILED_TO_PROCESS_IRIS_SHARES,
-                        );
-                        let serialized = serde_json::to_string(&message).unwrap();
-                        (reauth_error_result_attributes, serialized)
-                    }
-                    _ => unreachable!(), // we don't push a handle for unknown message types
-                };
-
-                send_error_results_to_sns(
-                    message,
-                    &batch_query.metadata[index],
-                    sns_client,
-                    config,
-                    result_attributes,
-                    batch_query.request_types[index].as_str(),
-                )
-                .await?;
-                // If we failed to process the iris shares, we include a dummy entry in the
-                // batch in order to keep the same order across nodes
-                let dummy_code_share = GaloisRingIrisCodeShare::default_for_party(party_id);
-                let dummy_mask_share = GaloisRingTrimmedMaskCodeShare::default_for_party(party_id);
-                (
-                    (
-                        (
-                            dummy_code_share.clone(),
-                            dummy_mask_share.clone(),
-                            dummy_code_share.clone().all_rotations(),
-                            dummy_mask_share.clone().all_rotations(),
-                            dummy_code_share.clone().all_rotations(),
-                            dummy_mask_share.clone().all_rotations(),
-                        ),
-                        (
-                            dummy_code_share.clone(),
-                            dummy_mask_share.clone(),
-                            dummy_code_share.clone().all_rotations(),
-                            dummy_mask_share.clone().all_rotations(),
-                            dummy_code_share.clone().all_rotations(),
-                            dummy_mask_share.clone().all_rotations(),
-                        ),
-                    ),
-                    false,
-                )
-            }
-        };
-
-        batch_query.valid_entries.push(valid_entry);
-
-        batch_query
-            .left_iris_requests
-            .code
-            .push(store_iris_shares_left);
-        batch_query
-            .left_iris_requests
-            .mask
-            .push(store_mask_shares_left);
-        batch_query
-            .left_iris_rotated_requests
-            .code
-            .extend(db_iris_shares_left);
-        batch_query
-            .right_iris_rotated_requests
-            .mask
-            .extend(db_mask_shares_left);
-        batch_query
-            .left_iris_interpolated_requests
-            .code
-            .extend(iris_shares_left);
-        batch_query
-            .left_iris_interpolated_requests
-            .mask
-            .extend(mask_shares_left);
-
-        batch_query
-            .right_iris_requests
-            .code
-            .push(store_iris_shares_right);
-        batch_query
-            .right_iris_requests
-            .mask
-            .push(store_mask_shares_right);
-        batch_query
-            .right_iris_rotated_requests
-            .code
-            .extend(db_iris_shares_right);
-        batch_query
-            .right_iris_rotated_requests
-            .mask
-            .extend(db_mask_shares_right);
-        batch_query
-            .right_iris_interpolated_requests
-            .code
-            .extend(iris_shares_right);
-        batch_query
-            .right_iris_interpolated_requests
-            .mask
-            .extend(mask_shares_right);
-    }
-
-    tracing::info!(
-        "Batch requests: {:?}",
-        batch_query
-            .request_ids
-            .iter()
-            .zip(batch_query.request_types.iter())
-            .collect::<Vec<_>>()
-    );
-
-    Ok(Some(batch_query))
-}
-
-fn get_iris_shares_parse_task(
-    party_id: usize,
-    shares_encryption_key_pairs: SharesEncryptionKeyPairs,
-    semaphore: Arc<Semaphore>,
-    s3_client_arc: S3Client,
-    bucket_name: String,
-    s3_key: String,
-) -> Result<JoinHandle<ParseSharesTaskResult>, ReceiveRequestError> {
-    let handle =
-        tokio::spawn(async move {
-            let _ = semaphore.acquire().await?;
-
-            let (share_b64, hash) =
-                match get_iris_data_by_party_id(&s3_key, party_id, &bucket_name, &s3_client_arc)
-                    .await
-                {
-                    Ok(iris_message_share) => iris_message_share,
-                    Err(e) => {
-                        tracing::error!("Failed to get iris shares: {:?}", e);
-                        eyre::bail!("Failed to get iris shares: {:?}", e);
-                    }
-                };
-
-            let iris_message_share =
-                match decrypt_iris_share(share_b64, shares_encryption_key_pairs.clone()) {
-                    Ok(iris_data) => iris_data,
-                    Err(e) => {
-                        tracing::error!("Failed to decrypt iris shares: {:?}", e);
-                        eyre::bail!("Failed to decrypt iris shares: {:?}", e);
-                    }
-                };
-
-            match validate_iris_share(hash, iris_message_share.clone()) {
-                Ok(_) => {}
-                Err(e) => {
-                    tracing::error!("Failed to validate iris shares: {:?}", e);
-                    eyre::bail!("Failed to validate iris shares: {:?}", e);
-                }
-            }
-
-            let (left_code, left_mask) = decode_iris_message_shares(
-                iris_message_share.left_iris_code_shares,
-                iris_message_share.left_mask_code_shares,
-            )?;
-
-            let (right_code, right_mask) = decode_iris_message_shares(
-                iris_message_share.right_iris_code_shares,
-                iris_message_share.right_mask_code_shares,
-            )?;
-
-            // Preprocess shares for left eye.
-            let left_future =
-                spawn_blocking(move || preprocess_iris_message_shares(left_code, left_mask));
-
-            // Preprocess shares for right eye.
-            let right_future =
-                spawn_blocking(move || preprocess_iris_message_shares(right_code, right_mask));
-
-            let (left_result, right_result) = tokio::join!(left_future, right_future);
-
-            Ok((
-                left_result.context("while processing left iris shares")??,
-                right_result.context("while processing right iris shares")??,
-            ))
-        });
-    Ok(handle)
-}
-
-#[tokio::main]
-async fn main() -> eyre::Result<()> {
-    dotenvy::dotenv().ok();
-
-    println!("Init config");
-    let mut config: Config = Config::load_config("SMPC").unwrap();
-    config.overwrite_defaults_with_cli_args(Opt::parse());
-
-    println!("Init tracing");
-    let _tracing_shutdown_handle = match initialize_tracing(&config) {
-        Ok(handle) => handle,
-        Err(e) => {
-            eprintln!("Failed to initialize tracing: {:?}", e);
-            return Err(e);
-        }
-    };
-
-    match server_main(config).await {
-        Ok(_) => {
-            tracing::info!("Server exited normally");
-        }
-        Err(e) => {
-            tracing::error!("Server exited with error: {:?}", e);
-            return Err(e);
-        }
-    }
-    Ok(())
-}
-
-async fn server_main(config: Config) -> eyre::Result<()> {
+pub async fn server_main(config: Config) -> eyre::Result<()> {
     let shutdown_handler = Arc::new(ShutdownHandler::new(
         config.shutdown_last_results_sync_timeout_secs,
     ));
@@ -1513,7 +860,7 @@ async fn server_main(config: Config) -> eyre::Result<()> {
 
     // Set the readiness flag to true, which will make the readiness server return a
     // 200 status code.
-    is_ready_flag_cloned.store(true, std::sync::atomic::Ordering::SeqCst);
+    is_ready_flag_cloned.store(true, Ordering::SeqCst);
 
     // Check other nodes and wait until all nodes are ready.
     let all_readiness_addresses = get_check_addresses(
@@ -1696,263 +1043,4 @@ async fn server_main(config: Config) -> eyre::Result<()> {
         }
     }
     Ok(())
-}
-
-// Helper function to load Aurora db records from the stream into memory
-async fn load_db_records<'a>(
-    actor: &mut impl InMemoryStore,
-    mut record_counter: i32,
-    all_serial_ids: &mut HashSet<i64>,
-    mut stream_db: BoxStream<'a, eyre::Result<DbStoredIris>>,
-) {
-    let mut load_summary_ts = Instant::now();
-    let mut time_waiting_for_stream = Duration::from_secs(0);
-    let mut time_loading_into_memory = Duration::from_secs(0);
-    let n_loaded_via_s3 = record_counter;
-    while let Some(iris) = stream_db.next().await {
-        // Update time waiting for the stream
-        time_waiting_for_stream += load_summary_ts.elapsed();
-        load_summary_ts = Instant::now();
-
-        let iris = iris.unwrap();
-
-        actor.load_single_record_from_db(
-            iris.index() - 1,
-            iris.left_code(),
-            iris.left_mask(),
-            iris.right_code(),
-            iris.right_mask(),
-        );
-
-        // Only increment db size if record has not been loaded via s3 before
-        if all_serial_ids.contains(&(iris.index() as i64)) {
-            actor.increment_db_size(iris.index() - 1);
-            all_serial_ids.remove(&(iris.index() as i64));
-            record_counter += 1;
-        }
-
-        // Update time spent loading into memory
-        time_loading_into_memory += load_summary_ts.elapsed();
-        load_summary_ts = Instant::now();
-    }
-
-    tracing::info!(
-        "Aurora Loading summary => Loaded {:?} items. Waited for stream: {:?}, Loaded into \
-         memory: {:?}",
-        record_counter - n_loaded_via_s3,
-        time_waiting_for_stream,
-        time_loading_into_memory,
-    );
-}
-
-async fn process_identity_deletions(
-    batch: &BatchQuery,
-    store: &Store,
-    dummy_iris_share: &GaloisRingIrisCodeShare,
-    dummy_mask_share: &GaloisRingTrimmedMaskCodeShare,
-) -> eyre::Result<()> {
-    if batch.deletion_requests_indices.is_empty() {
-        return Ok(());
-    }
-
-    for (&entry_idx, tracing_payload) in batch
-        .deletion_requests_indices
-        .iter()
-        .zip(batch.deletion_requests_metadata.iter())
-    {
-        let serial_id = entry_idx + 1; // DB serial_id is 1-indexed
-        tracing::info!(
-            node_id = tracing_payload.node_id,
-            dd.trace_id = tracing_payload.trace_id,
-            dd.span_id = tracing_payload.span_id,
-            "Started processing deletion request",
-        );
-
-        // overwrite postgres db with dummy values.
-        // note that both serial_id and postgres db are 1-indexed.
-        store
-            .update_iris(
-                None,
-                serial_id as i64,
-                dummy_iris_share,
-                dummy_mask_share,
-                dummy_iris_share,
-                dummy_mask_share,
-            )
-            .await?;
-
-        tracing::info!(
-            node_id = tracing_payload.node_id,
-            dd.trace_id = tracing_payload.trace_id,
-            dd.span_id = tracing_payload.span_id,
-            "Deleted identity with serial id {}",
-            serial_id,
-        );
-    }
-
-    Ok(())
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn load_db(
-    actor: &mut impl InMemoryStore,
-    store: &Store,
-    store_len: usize,
-    store_load_parallelism: usize,
-    config: &Config,
-    db_chunks_s3_store: impl ObjectStore,
-    db_chunks_s3_client: S3Client,
-    s3_chunks_folder_name: String,
-    s3_chunks_bucket_name: String,
-    s3_load_parallelism: usize,
-    s3_load_max_retries: usize,
-    s3_load_initial_backoff_ms: u64,
-    download_shutdown_handler: Arc<ShutdownHandler>,
-) -> eyre::Result<()> {
-    let total_load_time = Instant::now();
-    let now = Instant::now();
-
-    let mut record_counter = 0;
-    let mut all_serial_ids: HashSet<i64> = HashSet::from_iter(1..=(store_len as i64));
-    actor.reserve(store_len);
-
-    if config.enable_s3_importer {
-        tracing::info!("S3 importer enabled. Fetching from s3 + db");
-
-        // First fetch last snapshot from S3
-        let last_snapshot_details =
-            last_snapshot_timestamp(&db_chunks_s3_store, s3_chunks_folder_name.clone()).await?;
-
-        let min_last_modified_at =
-            last_snapshot_details.timestamp - config.db_load_safety_overlap_seconds;
-        tracing::info!(
-            "Last snapshot timestamp: {}, min_last_modified_at: {}",
-            last_snapshot_details.timestamp,
-            min_last_modified_at
-        );
-
-        let s3_store = S3Store::new(db_chunks_s3_client, s3_chunks_bucket_name);
-        let s3_arc = Arc::new(s3_store);
-
-        let (tx, mut rx) = mpsc::channel::<S3StoredIris>(config.load_chunks_buffer_size);
-
-        tokio::spawn(async move {
-            fetch_and_parse_chunks(
-                s3_arc,
-                s3_load_parallelism,
-                s3_chunks_folder_name,
-                last_snapshot_details,
-                tx.clone(),
-                s3_load_max_retries,
-                s3_load_initial_backoff_ms,
-            )
-            .await
-            .expect("Couldn't fetch and parse chunks from s3");
-        });
-
-        let mut time_waiting_for_stream = Duration::from_secs(0);
-        let mut time_loading_into_memory = Duration::from_secs(0);
-        let mut load_summary_ts = Instant::now();
-        while let Some(iris) = rx.recv().await {
-            time_waiting_for_stream += load_summary_ts.elapsed();
-            load_summary_ts = Instant::now();
-            let index = iris.index();
-
-            if index == 0 {
-                tracing::error!("Invalid iris index {}", index);
-                return Err(eyre!("Invalid iris index {}", index));
-            } else if index > store_len {
-                tracing::warn!(
-                    "Skip loading rolled back item: index {} > store_len {}",
-                    index,
-                    store_len
-                );
-                continue;
-            } else if !all_serial_ids.contains(&(index as i64)) {
-                tracing::warn!("Skip loading s3 retried item: index {}", index);
-                continue;
-            }
-
-            actor.load_single_record_from_s3(
-                iris.index() - 1,
-                iris.left_code_odd(),
-                iris.left_code_even(),
-                iris.right_code_odd(),
-                iris.right_code_even(),
-                iris.left_mask_odd(),
-                iris.left_mask_even(),
-                iris.right_mask_odd(),
-                iris.right_mask_even(),
-            );
-            actor.increment_db_size(index - 1);
-
-            if record_counter % 100_000 == 0 {
-                let elapsed = now.elapsed();
-                tracing::info!(
-                    "Loaded {} records into memory in {:?} ({:.2} entries/s)",
-                    record_counter,
-                    elapsed,
-                    record_counter as f64 / elapsed.as_secs_f64()
-                );
-                if download_shutdown_handler.is_shutting_down() {
-                    tracing::warn!("Shutdown requested by shutdown_handler.");
-                    return Err(eyre::eyre!("Shutdown requested"));
-                }
-            }
-
-            time_loading_into_memory += load_summary_ts.elapsed();
-            load_summary_ts = Instant::now();
-
-            all_serial_ids.remove(&(index as i64));
-            record_counter += 1;
-        }
-        tracing::info!(
-            "S3 Loading summary => Loaded {:?} items. Waited for stream: {:?}, Loaded into \
-             memory: {:?}.",
-            record_counter,
-            time_waiting_for_stream,
-            time_loading_into_memory,
-        );
-
-        let stream_db = store
-            .stream_irises_par(Some(min_last_modified_at), store_load_parallelism)
-            .await
-            .boxed();
-        load_db_records(actor, record_counter, &mut all_serial_ids, stream_db).await;
-    } else {
-        tracing::info!("S3 importer disabled. Fetching only from db");
-        let stream_db = store
-            .stream_irises_par(None, store_load_parallelism)
-            .await
-            .boxed();
-        load_db_records(actor, record_counter, &mut all_serial_ids, stream_db).await;
-    }
-
-    if !all_serial_ids.is_empty() {
-        tracing::error!("Not all serial_ids were loaded: {:?}", all_serial_ids);
-        return Err(eyre!(
-            "Not all serial_ids were loaded: {:?}",
-            all_serial_ids
-        ));
-    }
-
-    tracing::info!("Preprocessing db");
-    actor.preprocess_db();
-
-    tracing::info!(
-        "Loaded {} records from db into memory in {:?} [DB sizes: {:?}]",
-        record_counter,
-        total_load_time.elapsed(),
-        actor.current_db_sizes()
-    );
-
-    eyre::Ok(())
-}
-
-fn get_check_addresses(hostnames: Vec<String>, ports: Vec<String>, endpoint: &str) -> Vec<String> {
-    hostnames
-        .iter()
-        .zip(ports.iter())
-        .map(|(host, port)| format!("http://{}:{}/{}", host, port, endpoint))
-        .collect::<Vec<String>>()
 }
