+++ iris-mpc/src/services/aws/clients.rs
@@ -1,4 +1,4 @@
-use crate::aws::s3;
+use crate::services::aws::s3::{create_db_chunks_s3_client, create_s3_client};
 use aws_sdk_s3::Client as S3Client;
 use aws_sdk_secretsmanager::Client as SecretsManagerClient;
 use aws_sdk_sns::Client as SNSClient;
@@ -32,8 +32,8 @@ impl AwsClients {
 
         let force_path_style = config.environment != "prod" && config.environment != "stage";
 
-        let s3_client = s3::create_s3_client(&shared_config, force_path_style);
-        let db_chunks_s3_client = s3::create_db_chunks_s3_client(&shared_config, force_path_style);
+        let s3_client = create_s3_client(&shared_config, force_path_style);
+        let db_chunks_s3_client = create_db_chunks_s3_client(&shared_config, force_path_style);
         let secrets_manager_client = SecretsManagerClient::new(&shared_config);
 
         Ok(Self {
